requirements.txt
gradio==4.44.0
tensorflow-cpu==2.18.0
numpy==1.26.4
Pillow==10.4.0
requests==2.32.3

